17.2: Convert datatype to use Int128

+++ day_17/spec/test_operand.cr
@@ -3,7 +3,7 @@ class TestOperand < Operand
     new(raw_operand: value, context: Context.new)
   end
 
-  def value : Int32
+  def value : Int128
     @raw_operand
   end
 end
